extract FileRegister, FileRegisterManually, FileAccessManually components from App
